feat(contact-tagging): Backend: Moved loadEntityTags method to TagTransformer
