+++ apps/backend/src/api/transformers/CalloutResponseTransformer.ts
@@ -21,7 +21,7 @@ import { BaseCalloutResponseTransformer } from "@api/transformers/BaseCalloutRes
 import CalloutTransformer from "@api/transformers/CalloutTransformer";
 import CalloutResponseCommentTransformer from "@api/transformers/CalloutResponseCommentTransformer";
 import { calloutTagTransformer } from "@api/transformers/TagTransformer";
-import { batchUpdate, loadEntityTags } from "@api/utils";
+import { batchUpdate } from "@api/utils";
 
 import {
   Callout,
@@ -131,7 +131,11 @@ export class CalloutResponseTransformer extends BaseCalloutResponseTransformer<
       // TODO: Same logic as in ContactTransformer
       if (query.with?.includes(GetCalloutResponseWith.Tags)) {
         // Load tags after to ensure offset/limit work
-        await loadEntityTags(responses, CalloutResponseTag, "responseId");
+        await calloutTagTransformer.loadEntityTags(
+          responses,
+          CalloutResponseTag,
+          "responseId"
+        );
       }
     }
   }
