+++ apps/backend/src/api/utils/index.ts
@@ -1,5 +1,4 @@
 export * from "./auth.js";
 export * from "./objects.js";
 export * from "./rules.js";
-export * from "./tags.js";
 export * from "./validation.js";
